fix(navbar): use react-query for update avatar to solve the problem of avatar not updating
