+++ components/navbar/avatar.tsx
@@ -1,6 +1,12 @@
+"use client";
+
+import Image from "next/image";
+import { useEffect, useState } from "react";
+
 import NavAvatarLink from "./avatar-link";
 import LoginButton from "./login-button";
 
+import useAuth from "@/hooks/auth/useAuth";
 import { sessionType } from "@/types";
 
 export default function NavAvatar({
@@ -8,16 +14,21 @@ export default function NavAvatar({
 }: {
   initialSession: sessionType | null;
 }) {
-  const session = initialSession;
   const defaultAvatar = "/default-avatar.png";
+  const [session, setSession] = useState<sessionType | null>(initialSession);
+  const { session: newSession, isLoading } = useAuth();
 
-  if (session === null || session === undefined) return <LoginButton />;
+  useEffect(() => {
+    if (!isLoading && newSession !== undefined) {
+      setSession(newSession);
+    }
+  }, [isLoading]);
+
+  if (!session) return <LoginButton />;
 
   return (
     <NavAvatarLink>
-      {/* disable Next.js image optimize. it will cache the picture! */}
-      {/* eslint-disable-next-line @next/next/no-img-element*/}
-      <img
+      <Image
         alt={`${session.name}'s avatar`}
         className="rounded-full"
         height={"40"}
