+++ hooks/auth/useAuth.ts
@@ -3,9 +3,26 @@
 import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import {} from "jose";
 
-import { fetchSession, logout } from "@/lib/auth";
+import { fc } from "@/lib/fetchClient";
+import { sessionType } from "@/types";
 
 const AUTH_QUERY_KEY = ["auth", "session"];
+
+export async function fetchSession(): Promise<sessionType | null> {
+  try {
+    const session = await fc.get<sessionType>("/api/auth/me", {
+      cache: "no-store",
+      headers: { "x-timestamp": Date.now().toString() },
+    });
+
+    return session;
+  } catch (e) {
+    // eslint-disable-next-line no-console
+    console.error(e);
+
+    return null;
+  }
+}
 
 export default function useAuth() {
   const queryClient = useQueryClient();
@@ -22,7 +39,7 @@ export default function useAuth() {
   });
 
   const logoutMutation = useMutation({
-    mutationFn: logout,
+    mutationFn: async () => await fc.post("/api/auth/logout"),
     onSuccess: () => {
       queryClient.clear();
       queryClient.setQueryData(AUTH_QUERY_KEY, null);
