+++ lib/auth/verify-auth.ts
@@ -3,8 +3,6 @@
 import { jwtVerify } from "jose";
 import { cookies } from "next/headers";
 
-import { fc } from "../fetchClient";
-
 import { sessionType } from "@/types";
 
 export async function getSession(): Promise<sessionType | null> {
@@ -22,19 +20,6 @@ export async function getSession(): Promise<sessionType | null> {
   } catch {
     (await cookies()).delete("token");
 
-    return null;
-  }
-}
-
-export async function fetchSession(): Promise<sessionType | null> {
-  try {
-    const session = await fc.get<sessionType>("/api/auth/me", {
-      cache: "no-store",
-      headers: { "x-timestamp": Date.now().toString() },
-    });
-
-    return session;
-  } catch {
     return null;
   }
 }
