+++ middleware/auth.ts
@@ -1,3 +1,5 @@
+"use server";
+
 import { NextRequest, NextResponse } from "next/server";
 
 import { getSession } from "@/lib/auth";
