+++ package.json
@@ -32,8 +32,8 @@
     "@next/mdx": "^15.2.4",
     "@react-aria/ssr": "^3.9.7",
     "@react-aria/visually-hidden": "^3.8.21",
-    "@tanstack/react-query": "^5.71.1",
-    "@tanstack/react-query-devtools": "^5.71.2",
+    "@tanstack/react-query": "^5.71.3",
+    "@tanstack/react-query-devtools": "^5.71.3",
     "@types/mdx": "^2.0.13",
     "@uiw/react-md-editor": "^4.0.5",
     "bowser": "^2.11.0",
@@ -47,7 +47,7 @@
     "katex": "^0.16.21",
     "lowlight": "^3.3.0",
     "lxgw-wenkai-webfont": "^1.7.0",
-    "motion": "^12.6.2",
+    "motion": "^12.6.3",
     "next": "^15.2.4",
     "next-themes": "^0.4.6",
     "nextjs-toploader": "^3.8.16",
@@ -67,7 +67,7 @@
     "remark-gfm": "^4.0.1",
     "remark-math": "^6.0.0",
     "remark-parse": "^11.0.0",
-    "remark-rehype": "^11.1.1",
+    "remark-rehype": "^11.1.2",
     "remark-toc": "^9.0.0",
     "sharp": "^0.33.5",
     "unified": "^11.0.5",
@@ -80,11 +80,11 @@
     "@commitlint/config-conventional": "^19.8.0",
     "@next/eslint-plugin-next": "^15.2.4",
     "@react-types/shared": "^3.28.0",
-    "@tailwindcss/postcss": "^4.0.17",
+    "@tailwindcss/postcss": "^4.1.1",
     "@tailwindcss/typography": "^0.5.16",
-    "@types/node": "^22.13.16",
-    "@types/react": "^19.0.12",
-    "@types/react-dom": "^19.0.4",
+    "@types/node": "^22.13.17",
+    "@types/react": "^19.1.0",
+    "@types/react-dom": "^19.1.1",
     "@typescript-eslint/eslint-plugin": "^8.29.0",
     "@typescript-eslint/parser": "^8.29.0",
     "autoprefixer": "^10.4.21",
@@ -103,12 +103,12 @@
     "prettier": "^3.5.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
     "tailwind-variants": "^0.3.1",
-    "tailwindcss": "^4.0.17",
+    "tailwindcss": "^4.1.1",
     "tsx": "^4.19.3",
     "typescript": "^5.8.2"
   },
   "optionalDependencies": {
-    "wrangler": "^4.6.0",
-    "@opennextjs/cloudflare": "^0.6.5"
+    "@opennextjs/cloudflare": "^0.6.6",
+    "wrangler": "^4.6.0"
   }
 }
